refactor(generator): relocate routing from descriptor_utils to separate file
