+++ generator/CMakeLists.txt
@@ -100,6 +100,8 @@ add_library(
     internal/retry_traits_generator.h
     internal/round_robin_decorator_generator.cc
     internal/round_robin_decorator_generator.h
+    internal/routing.cc
+    internal/routing.h
     internal/sample_generator.cc
     internal/sample_generator.h
     internal/scaffold_generator.cc
@@ -222,6 +224,7 @@ function (google_cloud_cpp_generator_define_tests)
         internal/printer_test.cc
         internal/resolve_comment_references_test.cc
         internal/resolve_method_return_test.cc
+        internal/routing_test.cc
         internal/scaffold_generator_test.cc
         internal/service_code_generator_test.cc)
