+++ generator/google_cloud_cpp_generator.bzl
@@ -56,6 +56,7 @@ google_cloud_cpp_generator_hdrs = [
     "internal/resolve_method_return.h",
     "internal/retry_traits_generator.h",
     "internal/round_robin_decorator_generator.h",
+    "internal/routing.h",
     "internal/sample_generator.h",
     "internal/scaffold_generator.h",
     "internal/service_code_generator.h",
@@ -104,6 +105,7 @@ google_cloud_cpp_generator_srcs = [
     "internal/resolve_method_return.cc",
     "internal/retry_traits_generator.cc",
     "internal/round_robin_decorator_generator.cc",
+    "internal/routing.cc",
     "internal/sample_generator.cc",
     "internal/scaffold_generator.cc",
     "internal/service_code_generator.cc",
