+++ generator/google_cloud_cpp_generator_unit_tests.bzl
@@ -32,6 +32,7 @@ google_cloud_cpp_generator_unit_tests = [
     "internal/printer_test.cc",
     "internal/resolve_comment_references_test.cc",
     "internal/resolve_method_return_test.cc",
+    "internal/routing_test.cc",
     "internal/scaffold_generator_test.cc",
     "internal/service_code_generator_test.cc",
 ]
