+++ generator/internal/metadata_decorator_rest_generator.cc
@@ -17,6 +17,7 @@
 #include "generator/internal/http_option_utils.h"
 #include "generator/internal/predicate_utils.h"
 #include "generator/internal/printer.h"
+#include "generator/internal/routing.h"
 #include "google/cloud/internal/absl_str_cat_quiet.h"
 #include "absl/strings/str_split.h"
 #include <google/protobuf/descriptor.h>
